script: Move encoding interfaces into `encoding/`
Part of #38901
