+++ components/script/dom/encoding/mod.rs
@@ -1,0 +1,9 @@
+/* This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at https://mozilla.org/MPL/2.0/. */
+
+pub(crate) mod textdecoder;
+pub(crate) mod textdecodercommon;
+pub(crate) mod textdecoderstream;
+pub(crate) mod textencoder;
+pub(crate) mod textencoderstream;
