+++ components/script/dom/mod.rs
@@ -282,6 +282,8 @@ pub(crate) mod domtokenlist;
 #[expect(dead_code)]
 pub(crate) mod element;
 pub(crate) mod elementinternals;
+pub(crate) mod encoding;
+pub(crate) use self::encoding::*;
 pub(crate) mod errorevent;
 pub(crate) mod event;
 pub(crate) mod eventsource;
@@ -394,11 +396,6 @@ mod testing;
 pub(crate) use self::testing::*;
 pub(crate) mod text;
 pub(crate) mod textcontrol;
-pub(crate) mod textdecoder;
-pub(crate) mod textdecodercommon;
-pub(crate) mod textdecoderstream;
-pub(crate) mod textencoder;
-pub(crate) mod textencoderstream;
 pub(crate) mod timeranges;
 pub(crate) mod toggleevent;
 pub(crate) mod touch;
