+++ components/script/dom/stream/transformstreamdefaultcontroller.rs
@@ -28,12 +28,14 @@ use crate::dom::compressionstream::{
 use crate::dom::decompressionstream::{
     decompress_and_enqueue_a_chunk, decompress_flush_and_enqueue,
 };
+use crate::dom::encoding::textdecodercommon::TextDecoderCommon;
+use crate::dom::encoding::textdecoderstream::{decode_and_enqueue_a_chunk, flush_and_enqueue};
+use crate::dom::encoding::textencoderstream::{
+    Encoder, encode_and_enqueue_a_chunk, encode_and_flush,
+};
 use crate::dom::globalscope::GlobalScope;
 use crate::dom::promise::Promise;
 use crate::dom::promisenativehandler::{Callback, PromiseNativeHandler};
-use crate::dom::textdecodercommon::TextDecoderCommon;
-use crate::dom::textdecoderstream::{decode_and_enqueue_a_chunk, flush_and_enqueue};
-use crate::dom::textencoderstream::{Encoder, encode_and_enqueue_a_chunk, encode_and_flush};
 use crate::dom::types::{DecompressionStream, TransformStream};
 use crate::realms::{InRealm, enter_realm};
 use crate::script_runtime::{CanGc, JSContext as SafeJSContext};
